@@ -1,21 +1,21 @@
 name: zakupyapp
 description: Lista zakupowa
-version: 1.4.0  # updated 30/07/2022
+version: 1.5.0  # updated 06/09/2023
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  firebase_core: ^1.20.0
-  firebase_database: ^9.0.20
-  firebase_storage: ^10.3.4
-  shared_preferences: ^2.0.6
-  package_info_plus: ^1.4.2
-  test: ^1.21.1
-  url_launcher: ^6.1.5
-  firebase_app_check: ^0.0.6+18
+  firebase_core: ^2.15.1
+  firebase_database: ^10.2.5
+  firebase_storage: ^11.2.6
+  firebase_app_check: ^0.1.5+2
+  shared_preferences: ^2.2.1
+  package_info_plus: ^4.1.0
+  test: ^1.24.3
+  url_launcher: ^6.1.14
 
 dev_dependencies:
   flutter_test:
