@@ -1,1 +1,1 @@
-enum Urgency {too_late, urgent, not_urgent}
+enum Urgency { too_late, urgent, not_urgent }
