@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:zakupyapk/storage/storage_manager.dart';
+
+import 'package:zakupyapp/storage/storage_manager.dart';
 
 class SimpleTextWithIcon extends StatelessWidget {
   final String text;
