@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:url_launcher/url_launcher.dart';
-import 'package:zakupyapk/core/apprelease.dart';
-import 'package:zakupyapk/utils/app_info.dart';
+
+import 'package:zakupyapp/core/apprelease.dart';
+import 'package:zakupyapp/utils/app_info.dart';
 
 class DownloadUpdateDialog extends StatelessWidget {
   final AppRelease latestRelease;
